requirements.txt
aiogram
pydantic_settings
pydantic
ollama
loguru
opencv-python
pillow